Fix jlewi/foyle#188 (#1536)
https://github.com/jlewi/foyle/issues/188

The reason newly added ghost cells aren't executable is because they
lakc a runme.dev/id metadata. This field gets set when cells are inserted but it was being removed by the code that update the cell
to set ghost to false. This update field was replacing the entire metadata field and not just updating it.