Include output from interactive cells in Foyle requests (#1756)
* Include output from interactive cells in Foyle requests

This fixes a bug in the serializaiton of the notebook before sending
it to Foyle that caused the output of interactive cells not to be
included in the requests.

The problem is that we need to call addExecInfo before converting
the VSCode NotebookData representation to the proto. That
handles copying the output of the interactive terminals into
the NotebookData structure.

This necessitated some code refactoring. In order to call
addExecInfo we need an instance of the kernel.

We create a new Converter class to keep track of the kernel
and also provide reuse in the logic for converting notebook data to
protos for Foyle.

Since addExecInfo is async we need to change buildReq to return
a promise and refactor some of the logic to be non blocking.

* Fix jlewi/foyle#286

* Update to use await.

* Add a comment.

* Use await.